+++ src/drawBitmap.h
@@ -10,7 +10,7 @@ enum DrawMode: uint8_t {
     Xor
 };
 
-void drawBitmap(int16_t x, int16_t y, const uint8_t* bitmap, const uint8_t* mask, bool plusMask, uint8_t frame, MirrorMode mirror, DrawMode drawMode, uint8_t maskFrame = 255);
+void drawBitmap(int16_t x, int8_t y, const uint8_t* bitmap, const uint8_t* mask, bool plusMask, uint8_t frame, MirrorMode mirror, DrawMode drawMode, uint8_t maskFrame = 255);
 
 #endif
